Add short al
This is an Array list that is optimised for the case where there
may be a small number of items in each list. If the items take up
less space than the size of the pointers to allocated ones, then
they will be stored in the struct, rather than an allocation.

Unfortunately this is currently formulated in a way that need eagre
memory allocation and deallocation which could lead to poor
performance in edge cases.