Change hash table size to be actual size
There was a bug with the previous size system, as a size of 0 could
not be represented, unless the presence of the pointers was checked.

This is also a cleaner way of representing the size of the hash table,
and should have been done from the start.

The macros will be removed in a successive change.